Change Log: change in the way to release sessions in order to spare the dfc parameter dfc.compatibility.allow_weak_disconnect for DCTM6. Before this change, an exception DFC_SESSION_HANDLE_RESTRICTION was generated if this parameter was not set to true in dfc.properties.
------------------------------
* projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/DctmAuthorizationManager.java
change in authorizeDocids method : pass the session to the execute method of DmQuery, release the session from the releaseSessionAuto method of DmQuery

* projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/DctmTraversalManager.java
change in execQuery method : pass the sessions to the execute method of DmQuery, release the sessions from the releaseSessionAdd and releaseSessionAdd methods
 of DmQuery

* projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/DctmDocumentList.java
change in finalize method : release the sessions from the releaseSessionAdd and releaseSessionAdd methods of DmQuery

* projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/DctmSysObjectDocument.java
comments of the RepositoryDocumentException : not compatible with the connectorManager 1.1.0


* projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/dfcwrap/ISessionManager.java
creation of the attributes :  DfSessionDel,DfSessionAdd, DfSessionAuto
creation of the methods : setSessionDel, setSessionAdd, setSessionAuto, releaseSessionDel, releaseSessionAdd, releaseSessionAuto

* projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/dfcwrap/IQuery.java
creation of a new execute method whose parameters are of ISession and int types.

* projects/dctm-client/source/java/com/google/enterprise/connector/dctm/dctmdfcwrap/DmSessionManager.java
creation of the attributes :  DfSessionDel,DfSessionAdd, DfSessionAuto
creation of the methods : setSessionDel, setSessionAdd, setSessionAuto, releaseSessionDel, releaseSessionAdd, releaseSessionAuto

* projects/dctm-client/source/java/com/google/enterprise/connector/dctm/dctmdfcwrap/DmQuery.java
creation of a new execute method whose parameters are of ISession and int types.