Added missing files from the source and binary downloads.
The version.properties file is unused, but appears in the source
downloads. The RELEASE_NOTES were compiled from the individual release
notes in each release, and replace the version-specific release notes
in the projects directory. The translations match version 1.1.0, and
do not include translations of the advanced UI configuration strings.