Changed the nested table for the advanced configuration options to a tbody.
The nested table did not allow the labels and inputs to be aligned,
whereas with the tbody everything is part of the same table. Firefox,
Safari, and Chrome don't work correctly with display: block on a tbody
element, whereas IE does not support the table-row-group value. All of
the browsers support no value for the display property, so we don't
include the display property in the HTML when advanced configuration
is on, and we set style.display='' to turn it on.

M    projects/google-enterprise-connector-dctm/source/java/com/google/enterprise/connector/dctm/DctmConnectorType.java