Fix bug 7608960: Documentum Connector fails to handle large Authorization requests coming from GSA. The error message starts with:
    DM_QUERY_E_SYNTAX]error:  "A Parser Error (yacc stack overflow)

DQL has a limit of slightly less than 500 items in an IN list. This
change splits the list into lists of up to 400 items combined with OR.

There is no Ant infrastructure to run the live tests in dctm-client. I
hacked at them to confirm the bug, and that this fix is valid DQL
syntax. I need a quick fix for an onsite deployment engineer, so I'll
include the live test refactoring in a later CL.

Code review: https://codereview.appspot.com/6897044