Fix domain name matching in group lookup [b/13253159, 4 of 4]. Add a check for DQL matches against a subdomain.
The DQL query matches NetBIOS domains anywhere in the user DN (except the
TLD), and it matches DNS domains on a suffix. We need to make sure that in
both cases the matches are against the leftmost DC elements (RDNs) in the
user DN. This matching uses LdapName to ensure the correct parsing of the
DNs. For example, EXAMPLE and EXAMPLE.COM should match

    CN=Joseph,dc=example,dc=com

but not

    CN=Joseph,dc=ace,dc=example,dc=com

As noted in a comment, for brevity, we loop over all the query results in
case we want to log them, and we check for a user domain on each iteration.

Code review: https://codereview.appspot.com/72870046/