Fix repeated adds and deletes with different time zones (b/19484996).
DFC call to getTime() adjusts the time to local client time zone, and when this time stamp info is used in DQL for the checkpoint, it is not adjusted back to the time zone of server and this results in getting the same records again. 

Code changed to use "DATETOSTRING" function in DQL to get the time stamp info, which doesn't adjust for timezone offset.

Code changed in Checkpoint class to use dates as Strings instead of java Date, so that string format dates are used consistently eveywhere.

Made required changes to tests to run all the tests successfully. Need to add test cases.

Code review: https://codereview.appspot.com/205090043/