Use windows_domain property if user_ldap_dn does not have DC RDNs (b/19732599).
If the user if synchronized from LDAP, but the DN does not specify a
domain, then we can use the windows_domain property as we do for
unsynchronized users.

The implementation extracts the DC attributes from the DN as an
LdapName and checks isEmpty on the name.

Related changes:

* Extract first part of DctmAuthenticationManger.domainMatchesUser as
  IdentityUtil.getDomainComponents.

* Use getDomainComponents to implement getFirstDomainFromDN
  (effectively removing the second of two methods copied from secmgr).

Also handle additional corner cases:

* Lookup of dm_user by user_login_name fails and returns null
* user_source = LDAP but user_ldap_dn is null or empty

Other changes:

* Remove unused imports and fix copyright headers.

Code review: https://codereview.appspot.com/219020043/